Battery inspect string patch, hediff added fix
Put all of the battery patches into a single class, added a patch to remove the self-discharge readout from the inspect string if self-discharging is disabled. Fixed a bug in HediffComp_MultiplyOtherHediffSeverity that caused severity to be set to only integer values.

Confirmed fix for issue #9

Branch should be ready for merge code-wise, but more documentation is needed.
