+++ Source/communityframework/communityframework/Harmony patches/ShortCircuitUtility/ShortCircuitPatches.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Reflection;
 using System.Reflection.Emit;
 using Verse;
